frenchTarot: refactor board component
