+++ web/src/games/frenchTarot/board.module.css
